Move Processors out of the internal package
They are part of the public API of `RangesListFactory` now.

While at it, also move the only remaning `RangeProcessorPipeline` one
level up, out of the otherwise empty `internal.range` package.
