+++ src/main/java/org/semver4j/internal/RangeProcessorPipeline.java
@@ -1,7 +1,7 @@
-package org.semver4j.internal.range;
+package org.semver4j.internal;
 
 import org.jspecify.annotations.NullMarked;
-import org.semver4j.internal.range.processor.Processor;
+import org.semver4j.processor.Processor;
 
 import java.util.ArrayList;
 import java.util.List;
