+++ src/main/java/org/semver4j/processor/XRangeProcessor.java
@@ -1,4 +1,4 @@
-package org.semver4j.internal.range.processor;
+package org.semver4j.processor;
 
 import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.Nullable;
@@ -14,7 +14,7 @@ import static java.lang.String.join;
 import static java.util.regex.Pattern.compile;
 import static org.semver4j.Range.RangeOperator.*;
 import static org.semver4j.internal.Tokenizers.XRANGE;
-import static org.semver4j.internal.range.processor.RangesUtils.*;
+import static org.semver4j.processor.RangesUtils.*;
 
 /**
  * <p>Processor for translate <a href="https://github.com/npm/node-semver#x-ranges-12x-1x-12-">X-Ranges</a> into classic
